Completely revamp how bins are linked run
This was needed to make it possible to add multiple source directories. Useful for having 2 seprate directories of tools. For example - I have useful accounting tools that I can't share publicly, but other tools like `nodels` I'd be happy to share in a repository.

Overall, it seems like just a better approach in general, and I now have a foundation for a configuration file option as well.

This took way too long - I'm pooped, but there are still a few important things to do - to be continued. This is stable enough for now.